requirements.txt
mosquitto
paho-mqtt
python-mpd2
systemd-python
beautifulsoup4
bs4
